improve oop structure for gui, added auth for service layer interactions, made tickets show email instead of id for friendlness

+++ src/service/CategoryService.java
@@ -1,6 +1,9 @@
 package service;
 
 import dao.CategoryDAO;
+import model.Admin;
+import model.User;
+
 import java.util.List;
 
 public class CategoryService {
@@ -15,11 +18,23 @@ public class CategoryService {
         return categoryDAO.getCategories();
     }
 
-    public void addCategory(String category) {
+    public void addCategory(User currentUser, String category) {
+
+        if (!(currentUser instanceof Admin)) {
+            throw new SecurityException(
+                    "Only administrators can add categories.");
+        }
+
         categoryDAO.addCategory(category);
     }
 
-    public void deleteCategory(String category) {
+    public void deleteCategory(User currentUser, String category) {
+
+        if (!(currentUser instanceof Admin)) {
+            throw new SecurityException(
+                    "Only administrators can delete categories.");
+        }
+
         categoryDAO.deleteCategory(category);
     }
 }
